update ClarNotationShapeFileAttribute to match hlnug requirements
