+++ triturus/src/main/java/org/n52/v3d/triturus/geologic/examples/GocadFileInfoApp.java
@@ -47,7 +47,7 @@ import org.n52.v3d.triturus.geologic.importers.IoGocadTSurfReader;
 public class GocadFileInfoApp
 {
 	private String
-		inFilename = "/projects/GeologicToolbox/data/simple_test.ts";
+		inFilename = "res/s_geologie_Rotliegend_ts.ts";
 	
 	public static void main(String args[]) {
 		new GocadFileInfoApp().run(args);
